Remove unneeded oc install information
By default oc 3.7.0 will be installed. However we are overwriting
this information with an out-of-date data structure containing information
such as the oc version and a checksum.. This is resulting
in any users who do not already have oc cached locally or installed
to experience an error with trying to access the data structure.

This change removes the overridden data structure so that the default
data structure can be used and install oc 3.7.0.