Ansible Galaxy Roles Updating
Currently when we install roles in the install script they will not be overwritten with updated roles. Instead the old role will remain.

This fix specifies the version in the requirements.yml, by just adding the version alone was not enough to update the roles, adding --force tag, forces the update if an update is required, deleting the old role and installing the updated role in its place.